Fix DAFootprint calculation in block header
