+++ turbo/jsonrpc/otterscan_generic_tracer.go
@@ -2,12 +2,12 @@ package jsonrpc
 
 import (
 	"context"
-	"github.com/erigontech/erigon-lib/opstack"
 
 	"github.com/erigontech/erigon-lib/chain"
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/kv"
 	"github.com/erigontech/erigon-lib/log/v3"
+	"github.com/erigontech/erigon-lib/opstack"
 	"github.com/erigontech/erigon/turbo/rpchelper"
 
 	"github.com/erigontech/erigon/core"
