+++ turbo/transactions/call.go
@@ -3,7 +3,6 @@ package transactions
 import (
 	"context"
 	"fmt"
-	"github.com/erigontech/erigon-lib/opstack"
 	"time"
 
 	"github.com/erigontech/erigon-lib/log/v3"
@@ -12,6 +11,7 @@ import (
 	"github.com/erigontech/erigon-lib/chain"
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/kv"
+	"github.com/erigontech/erigon-lib/opstack"
 	"github.com/erigontech/erigon/consensus"
 	"github.com/erigontech/erigon/core"
 	"github.com/erigontech/erigon/core/state"
